Fix check:types
