+++ packages/bbui/src/index.ts
@@ -4,25 +4,25 @@ import "./bbui.css"
 export * from "./constants"
 
 // Form components
-export { default as Checkbox } from "./Form/Checkbox.svelte"
-export { default as Combobox } from "./Form/Combobox.svelte"
-export { default as DatePicker } from "./Form/DatePicker.svelte"
-export { default as DateRangePicker } from "./Form/DateRangePicker.svelte"
-export { default as Dropzone } from "./Form/Dropzone.svelte"
-export { default as EnvDropdown } from "./Form/EnvDropdown.svelte"
-export { default as FieldLabel } from "./Form/FieldLabel.svelte"
-export { default as File } from "./Form/File.svelte"
 export { default as Input } from "./Form/Input.svelte"
-export { default as InputDropdown } from "./Form/InputDropdown.svelte"
-export { default as Multiselect } from "./Form/Multiselect.svelte"
-export { default as RadioGroup } from "./Form/RadioGroup.svelte"
-export { default as RichTextField } from "./Form/RichTextField.svelte"
-export { default as Search } from "./Form/Search.svelte"
-export { default as Select } from "./Form/Select.svelte"
-export { default as Slider } from "./Form/Slider.svelte"
 export { default as Stepper } from "./Form/Stepper.svelte"
 export { default as TextArea } from "./Form/TextArea.svelte"
+export { default as Select } from "./Form/Select.svelte"
+export { default as Combobox } from "./Form/Combobox.svelte"
+export { default as Dropzone } from "./Form/Dropzone.svelte"
+export { default as DatePicker } from "./Form/DatePicker.svelte"
+export { default as DateRangePicker } from "./Form/DateRangePicker.svelte"
 export { default as Toggle } from "./Form/Toggle.svelte"
+export { default as RadioGroup } from "./Form/RadioGroup.svelte"
+export { default as Checkbox } from "./Form/Checkbox.svelte"
+export { default as InputDropdown } from "./Form/InputDropdown.svelte"
+export { default as EnvDropdown } from "./Form/EnvDropdown.svelte"
+export { default as Multiselect } from "./Form/Multiselect.svelte"
+export { default as Search } from "./Form/Search.svelte"
+export { default as RichTextField } from "./Form/RichTextField.svelte"
+export { default as FieldLabel } from "./Form/FieldLabel.svelte"
+export { default as Slider } from "./Form/Slider.svelte"
+export { default as File } from "./Form/File.svelte"
 
 // Core form components to be used elsewhere (standard components)
 export * from "./Form/Core"
@@ -31,63 +31,64 @@ export * from "./Form/Core"
 export * from "./FancyForm"
 
 // Components
-export { default as Accordion } from "./Accordion/Accordion.svelte"
+export { default as Drawer } from "./Drawer/Drawer.svelte"
+export { default as DrawerContent } from "./Drawer/DrawerContent.svelte"
+export { default as Avatar } from "./Avatar/Avatar.svelte"
 export { default as ActionButton } from "./ActionButton/ActionButton.svelte"
 export { default as ActionGroup } from "./ActionGroup/ActionGroup.svelte"
 export { default as ActionMenu } from "./ActionMenu/ActionMenu.svelte"
-export { default as Avatar } from "./Avatar/Avatar.svelte"
-export { default as Badge } from "./Badge/Badge.svelte"
-export { default as Banner } from "./Banner/Banner.svelte"
-export { default as BannerDisplay } from "./Banner/BannerDisplay.svelte"
 export { default as Button } from "./Button/Button.svelte"
 export { default as ButtonGroup } from "./ButtonGroup/ButtonGroup.svelte"
 export { default as CollapsedButtonGroup } from "./ButtonGroup/CollapsedButtonGroup.svelte"
 export { default as ClearButton } from "./ClearButton/ClearButton.svelte"
-export { default as ColorPicker } from "./ColorPicker/ColorPicker.svelte"
-export { default as DetailSummary } from "./DetailSummary/DetailSummary.svelte"
-export { default as Divider } from "./Divider/Divider.svelte"
-export { default as Drawer } from "./Drawer/Drawer.svelte"
-export { default as DrawerContent } from "./Drawer/DrawerContent.svelte"
 export { default as Icon } from "./Icon/Icon.svelte"
 export { default as IconAvatar } from "./Icon/IconAvatar.svelte"
-export { default as IconPicker } from "./IconPicker/IconPicker.svelte"
-export { default as InlineAlert } from "./InlineAlert/InlineAlert.svelte"
-export { default as CopyInput } from "./Input/CopyInput.svelte"
+export { default as DetailSummary } from "./DetailSummary/DetailSummary.svelte"
+export { default as Popover, type PopoverAPI } from "./Popover/Popover.svelte"
+export { default as ProgressBar } from "./ProgressBar/ProgressBar.svelte"
+export { default as ProgressCircle } from "./ProgressCircle/ProgressCircle.svelte"
 export { default as Label } from "./Label/Label.svelte"
 export { default as Layout } from "./Layout/Layout.svelte"
 export { default as Page } from "./Layout/Page.svelte"
 export { default as Link } from "./Link/Link.svelte"
-export { default as List } from "./List/List.svelte"
-export { default as ListItem } from "./List/ListItem.svelte"
-export { default as MarkdownEditor } from "./Markdown/MarkdownEditor.svelte"
-export { default as MarkdownViewer } from "./Markdown/MarkdownViewer.svelte"
-export { default as MenuItem } from "./Menu/Item.svelte"
+export { default as Tooltip } from "./Tooltip/Tooltip.svelte"
+export { default as TempTooltip } from "./Tooltip/TempTooltip.svelte"
+export { default as TooltipWrapper } from "./Tooltip/TooltipWrapper.svelte"
+export { default as ContextTooltip } from "./Tooltip/Context.svelte"
 export { default as Menu } from "./Menu/Menu.svelte"
 export { default as MenuSection } from "./Menu/Section.svelte"
 export { default as MenuSeparator } from "./Menu/Separator.svelte"
+export { default as MenuItem } from "./Menu/Item.svelte"
 export { default as Modal, type ModalAPI } from "./Modal/Modal.svelte"
-export { keepOpen, default as ModalContent } from "./Modal/ModalContent.svelte"
-export { default as Notification } from "./Notification/Notification.svelte"
+export { default as ModalContent, keepOpen } from "./Modal/ModalContent.svelte"
 export { default as NotificationDisplay } from "./Notification/NotificationDisplay.svelte"
-export { default as Pagination } from "./Pagination/Pagination.svelte"
-export { default as PhosphorIconPicker } from "./PhosphorIconPicker/PhosphorIconPicker.svelte"
-export { default as Popover, type PopoverAPI } from "./Popover/Popover.svelte"
-export { default as ProgressBar } from "./ProgressBar/ProgressBar.svelte"
-export { default as ProgressCircle } from "./ProgressCircle/ProgressCircle.svelte"
-export { default as StatusLight } from "./StatusLight/StatusLight.svelte"
-export { default as Switcher } from "./Switcher/Switcher.svelte"
+export { default as Notification } from "./Notification/Notification.svelte"
+export { default as Context } from "./context"
 export { default as Table } from "./Table/Table.svelte"
-export { default as Tab } from "./Tabs/Tab.svelte"
 export { default as Tabs } from "./Tabs/Tabs.svelte"
-export { default as Tag } from "./Tags/Tag.svelte"
+export { default as Tab } from "./Tabs/Tab.svelte"
 export { default as Tags } from "./Tags/Tags.svelte"
-export { default as AbsTooltip } from "./Tooltip/AbsTooltip.svelte"
-export { default as ContextTooltip } from "./Tooltip/Context.svelte"
-export { default as TempTooltip } from "./Tooltip/TempTooltip.svelte"
-export { default as Tooltip } from "./Tooltip/Tooltip.svelte"
-export { default as TooltipWrapper } from "./Tooltip/TooltipWrapper.svelte"
-export { default as TreeItem } from "./TreeView/Item.svelte"
+export { default as Tag } from "./Tags/Tag.svelte"
 export { default as TreeView } from "./TreeView/Tree.svelte"
+export { default as TreeItem } from "./TreeView/Item.svelte"
+export { default as Divider } from "./Divider/Divider.svelte"
+export { default as Pagination } from "./Pagination/Pagination.svelte"
+export { default as Badge } from "./Badge/Badge.svelte"
+export { default as StatusLight } from "./StatusLight/StatusLight.svelte"
+export { default as ColorPicker } from "./ColorPicker/ColorPicker.svelte"
+export { default as IconPicker } from "./IconPicker/IconPicker.svelte"
+export { default as PhosphorIconPicker } from "./PhosphorIconPicker/PhosphorIconPicker.svelte"
+export { default as InlineAlert } from "./InlineAlert/InlineAlert.svelte"
+export { default as Banner } from "./Banner/Banner.svelte"
+export { default as CopyInput } from "./Input/CopyInput.svelte"
+export { default as BannerDisplay } from "./Banner/BannerDisplay.svelte"
+export { default as MarkdownEditor } from "./Markdown/MarkdownEditor.svelte"
+export { default as MarkdownViewer } from "./Markdown/MarkdownViewer.svelte"
+export { default as List } from "./List/List.svelte"
+export { default as ListItem } from "./List/ListItem.svelte"
+export { default as Accordion } from "./Accordion/Accordion.svelte"
+export { default as AbsTooltip } from "./Tooltip/AbsTooltip.svelte"
+export { default as Switcher } from "./Switcher/Switcher.svelte"
 
 // Renderers
 export { default as BoldRenderer } from "./Table/BoldRenderer.svelte"
@@ -96,18 +97,18 @@ export { default as InternalRenderer } from "./Table/InternalRenderer.svelte"
 
 // Typography
 export { default as Body } from "./Typography/Body.svelte"
-export { default as Code } from "./Typography/Code.svelte"
-export { default as Detail } from "./Typography/Detail.svelte"
 export { default as Heading } from "./Typography/Heading.svelte"
+export { default as Detail } from "./Typography/Detail.svelte"
+export { default as Code } from "./Typography/Code.svelte"
 
 // Actions
 export { default as autoResizeTextArea } from "./Actions/autoresize_textarea"
-export { default as clickOutside } from "./Actions/click_outside"
 export { default as positionDropdown } from "./Actions/position_dropdown"
+export { default as clickOutside } from "./Actions/click_outside"
 
 // Stores
+export { notifications, createNotificationStore } from "./Stores/notifications"
 export { banner, BANNER_TYPES } from "./Stores/banner"
-export { createNotificationStore, notifications } from "./Stores/notifications"
 
 // Helpers
 export * as Helpers from "./helpers"
